refactor: extract page components
