+++ app/contact/page.tsx
@@ -13,6 +13,7 @@ const ScrollArrow = lazy(() => import("../../components/layout/ScrollArrow"))
 const Cursor = lazy(() => import("../../components/layout/Cursor"))
 const LogoTitle = lazy(() => import("../../components/layout/LogoTitle"))
 const Footer = lazy(() => import("../../components/sections/Footer"))
+import { InputField, TextareaField, BackgroundAnimation } from "../../components/pages/contact"
 
 const ContactPage = () => {
   const [isSubmitting, setIsSubmitting] = useState(false)
@@ -198,66 +199,6 @@ const ContactPage = () => {
   )
 }
 
-const InputField = React.memo(({ type, name, placeholder, required }) => (
-  <motion.div className="relative overflow-hidden rounded-lg" whileHover={{ scale: 1.02 }} whileTap={{ scale: 0.98 }}>
-    <motion.input
-      type={type}
-      name={name}
-      placeholder={placeholder}
-      required={required}
-      className="w-full px-6 py-4 bg-white bg-opacity-5 text-white placeholder-white placeholder-opacity-50 focus:outline-none transition-colors z-10 relative"
-      whileFocus={{ scale: 1.02 }}
-    />
-    <motion.div
-      className="absolute inset-0 bg-gradient-to-r from-purple-500 to-pink-500 opacity-20"
-      initial={{ x: "-100%" }}
-      whileHover={{ x: "100%" }}
-      transition={{ duration: 0.5 }}
-    />
-  </motion.div>
-))
-InputField.displayName = "InputField"
-
-const TextareaField = React.memo(({ name, placeholder, required }) => (
-  <motion.div className="relative overflow-hidden rounded-lg" whileHover={{ scale: 1.02 }} whileTap={{ scale: 0.98 }}>
-    <motion.textarea
-      name={name}
-      placeholder={placeholder}
-      required={required}
-      className="w-full px-6 py-4 bg-white bg-opacity-5 text-white placeholder-white placeholder-opacity-50 focus:outline-none transition-colors resize-none h-40 z-10 relative"
-      whileFocus={{ scale: 1.02 }}
-    />
-    <motion.div
-      className="absolute inset-0 bg-gradient-to-r from-purple-500 to-pink-500 opacity-20"
-      initial={{ x: "-100%" }}
-      whileHover={{ x: "100%" }}
-      transition={{ duration: 0.5 }}
-    />
-  </motion.div>
-))
-TextareaField.displayName = "TextareaField"
-
-const BackgroundAnimation = React.memo(() => (
-  <motion.div
-    className="fixed inset-0 pointer-events-none z-10"
-    style={{
-      backgroundImage: "radial-gradient(circle, rgba(255,255,255,0.15) 1px, transparent 1px)",
-      backgroundSize: "30px 30px",
-    }}
-    animate={{
-      backgroundPosition: ["0px 0px", "0px -30px"],
-    }}
-    transition={{
-      backgroundPosition: {
-        repeat: Number.POSITIVE_INFINITY,
-        repeatType: "loop",
-        duration: 10,
-        ease: "linear",
-      },
-    }}
-  />
-))
-BackgroundAnimation.displayName = "BackgroundAnimation"
 
 export default ContactPage
