+++ components/pages/about-us/index.ts
@@ -1,0 +1,2 @@
+export { default as AboutSection } from "./AboutSection"
+export type { AboutSectionProps } from "./AboutSection"
