+++ components/pages/contact/index.ts
@@ -1,0 +1,3 @@
+export { default as InputField } from "./InputField"
+export { default as TextareaField } from "./TextareaField"
+export { default as BackgroundAnimation } from "./BackgroundAnimation"
