+++ components/pages/services/index.ts
@@ -1,0 +1,3 @@
+export { default as ServiceCard } from "./ServiceCard"
+export { default as ProcessStep } from "./ProcessStep"
+export { default as GlowingCard } from "./GlowingCard"
